Fixed an issue with statistics chart. Changed image model.
The empty chart was displayed for old threads. This sometimes occured because minimal post creation time was relative to the current time. Now it is relative to the creation time of the last returned post.

Post per hour in the statistics is now rounded to two decimal spaces, sometimes it was rounded down to 0.

Replaced Image.is_webm() with Image.get_extension(). The new method is more versatile.

ajax_gallery accepts an amount parameter now. It allows to specify the number of returned images.